requirements.txt
requests
django>=2
djangorestframework>=3.7.7
markdown>=2.6.7
django-filter>=0.15.3
django-nose
coverage
django-gm2m
djangorestframework-csv
pika
git+https://github.com/Semprini/drf-nest.git
git+https://github.com/Semprini/py-cdaf.git
